LayoutTests/webgl/2.0.0/conformance2/buffers/buffer-data-and-buffer-sub-data-sub-source-expected.txt

CONSOLE MESSAGE: line 55: ReferenceError: Can't find variable: iit
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/buffers/buffer-data-and-buffer-sub-data-sub-source.html
[ 1: PASS ] WebGL context exists
[ 2: PASS ] [object WebGLBuffer] is non-null.
[ 3: PASS ] getError was expected value: NO_ERROR :
[ 4: PASS ] getError was expected value: INVALID_VALUE : calling bufferData when srcOffset + length is larger than source size
[ 5: PASS ] getError was expected value: INVALID_VALUE : calling bufferData when srcOffset + length is larger than source size
[ 6: PASS ] getError was expected value: INVALID_VALUE : calling bufferData when srcOffset + length is larger than source size
[ 7: PASS ] getError was expected value: NO_ERROR : calling bufferData with valid sub source should succeed
[ 8: FAIL ] successfullyParsed should be true (of type boolean). Was undefined (of type undefined).
[ FAIL ] 1 failures reported
